Rename payout_curve to inverse
So that we can differentiate between different types of payout curves
once we introduce the second one.
