+++ model/src/payouts/inverse.rs
