+++ model/src/payouts/inverse/README.md
