+++ model/src/payouts/inverse/compat.rs
