+++ xtra-libp2p-rollover/src/v_1_0_0/protocol.rs
@@ -383,7 +383,7 @@ pub(crate) async fn build_own_cfd_transactions(
             id: announcement.id.to_string(),
             nonce_pks: announcement.nonce_pks.clone(),
         },
-        Payouts::new(
+        Payouts::new_inverse(
             our_position,
             punish_params.own_role,
             rollover_params.price,
